api/requirements.txt

colour-science
flask
flask-cors
numpy
networkx

